json data source (WIP)
This data source queries arbitrary endpoints that can return timeseries data.
The response must be JSON in the format expected by Grafana as the result of
the data source `query` method.

The configured URL should substitute `from` and `to` variables. The time format
used for those variables is configured in the data source settings.

Future work:
- Allow substitutions of more than just `from` and `to` and have those additional
  substitutions available in the metric editor.